+++ apps/meet/CHANGELOG.md
@@ -4,6 +4,4 @@
 
 ### Patch Changes
 
-- Updated dependencies [[`3cea6ee`](https://github.com/PulseBeamDev/pulsebeam-js/commit/3cea6ee0cd43828abc1c91d66dfcbd795ab4b69c)]:
-  - @pulsebeam/ui@0.1.0
-
+- Initial local React SDK package.